Fixed bug where Nodes wouldn't change values
For "nengo.Node([0,0])", the value array was being kept
as an int, causing it to round inputs to 0

+++ nengo_gui/javaviz/javaviz.py
@@ -169,9 +169,9 @@ class OverrideFunction(object):
         while OverrideFunction.overrides['block_time'] < t:
             time.sleep(0.01)
         if callable(self.function):
-            value = np.array(self.function(t))
+            value = np.array(self.function(t), dtype='float')
         else:
-            value = np.array(self.function)
+            value = np.array(self.function, dtype='float')
         for k,v in OverrideFunction.overrides.get(self.id, {}).items():
             value[k] = v
         return value
